+++ CHANGELOG.md
@@ -29,6 +29,8 @@ Other blocks are:
 ### Added
 - **OAR**: New options `--overlap` and `--overlap-limit` to print details on the overlap between reference and new sets of normal modes.
 - **LIB**(`visual.povrender`): Added POV-Ray `version` and `assumed_gamma` directives, as recommended since POV-Ray version 3.7.
+- **LIB**(`tools`): New module `tools.output` to provide simple functions for common printing/output-related operations.
+- **LIB**(`tools.output`): New function `fortran_fmt_D` to simulate old Fortran `D` format, primarily intended for quick comparisons with output produced by Fortran programs.
 - **API**: Added parsing of Y (Upsilon) matrix in VPT2 calculations from Gaussian log files.
 - **API**: Added parsing of first derivatives of electronic transitions dipole moments for electric dipole, magnetic dipole and electric quadrupole from Gaussian log files.